# 35% better performance on `scripts/benchmark_big_table.py`