--- a/htpy/__init__.py
+++ b/htpy/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+
 import dataclasses
 import functools
 import keyword
@@ -167,6 +168,7 @@ class Context(t.Generic[T]):
 
 def iter_node(x: Node) -> Iterator[str]:
     return _iter_node_context(x, {})
+
 
 def _iter_node_context(x: Node, context_dict: dict[Context[t.Any], t.Any]) -> Iterator[str]:
     while not isinstance(x, BaseElement) and callable(x):